pyproject.toml
[project]
name = "nolan-portfolio"
version = "0.1.0"
description = "Personal portfolio website for Nolan Fox"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-fasthtml>=0.6.0",
    "uvicorn>=0.30.0",
    "pyyaml>=6.0",
]

[project.scripts]
dev = "uvicorn main:app --reload"
